backend/requirements.txt
flask
flask-cors
pymongo
requests
beautifulsoup4
python-dotenv